+++ packages/orq-rc/src/models/errors/index.ts
@@ -6,3 +6,4 @@ export * from "./apierror.js";
 export * from "./honoapierror.js";
 export * from "./httpclienterrors.js";
 export * from "./sdkvalidationerror.js";
+export * from "./updateprompt.js";